+++ implementations/snmp/bin/snmp.py
@@ -1,12 +1,4 @@
-'''
-Modular Input Script
-
-Copyright (C) 2012 Splunk, Inc.
-All Rights Reserved
-
-'''
-
-import os,sys,logging
+import os,sys,logging,hashlib
 import xml.dom.minidom, xml.sax.saxutils
 import time
 import threading
@@ -57,6 +49,12 @@ SCHEME = """<scheme>
                 <title>SNMP Input Name</title>
                 <description>Name of this SNMP input</description>
             </arg>  
+            <arg name="activation_key">
+                <title>Activation Key</title>
+                <description>Visit http://www.baboonbones.com/#activation to obtain a free,non-expiring key</description>
+                <required_on_edit>true</required_on_edit>
+                <required_on_create>true</required_on_create>
+            </arg>
             <arg name="snmp_mode">
                 <title>SNMP Mode</title>
                 <description>Whether or not this stanza is for polling attributes or listening for traps</description>
@@ -361,6 +359,16 @@ def trapCallback(transportDispatcher, transportDomain, transportAddress, wholeMs
 def do_run():
     
     config = get_input_config() 
+    
+    activation_key = config.get("activation_key")
+    app_name = "SNMP Modular Input"
+    
+    m = hashlib.md5()
+    m.update((app_name))
+    if not m.hexdigest().upper() == activation_key.upper():
+        logging.error("FATAL Activation key for App '%s' failed" % app_name)
+        sys.exit(2)
+        
     #params
     snmp_mode=config.get("snmp_mode","")