+++ implementations/meraki/appServer/static/RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+0.8
+-----
+Added an activation key requirement , visit http://www.baboonbones.com/#activation to obtain a free,non-expiring key
+Docs updated
+Splunk 7.1 compatible
+
 0.7
 ---
 Moved the Meraki secret and validator out of inputs.conf. These values are now declared in your App's setup screen and the secret gets encrypted to app.conf and dynamically retreived by the meraki.py script