+++ implementations/tesla/appServer/static/RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+0.7
+-----
+Added an activation key requirement , visit http://www.baboonbones.com/#activation to obtain a free,non-expiring key
+Docs updated
+Splunk 7.1 compatible
+
 0.6
 ---
 Updated endpoints URLs to latest unofficial API